Make a cleaner implementation of downloading with `aria2c`.
This way, we are not hardcoding the cookies in text and, also, we are not
tied to youtube only anymore: the cookies are not specific to youtube
anymore.

Note to self: to make the cookies from this session persistent, we follow
what Alex Martelli says in [a post in Stack Overflow][0] to iterate from the
current cookie jar and set the cookies in another, file-backed jar.

[0]: http://stackoverflow.com/questions/1023224/how-to-pickle-a-cookiejar/1023235#1023235