InfoExtractors: facebook: Don't make an if without an else.
Under certain circumstances (relaxing previous constraints of the code), we
may find that the list of URLs gathered from a page is empty. This is tested
for in an `if`, but there is no corresponding guarding `else` to handle the
case of the list being empty.

This commit fixes this.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>